Instead of patching Buildroot, patch the rpi-userland package
This is achieved by imported the patch for the package under
patches/rpi-userland, and setting the BR2_GLOBAL_PATCH_DIR variable
pointing to the patches/ subdirectory. Buildroot will apply patches
from the global directory to each package, without needing to patch
the Buildroot main sources.

The ultimate goal is NOT needing to patch Buildroot at all, and
remove the top-level Makefile which downloads and patches the Buildroot
source, so the user of the overlay can use whichever Buildroot checkout
(or release tarball) as they desire.